Add `podspec`s for CocoaPods
This patch adds two `podspec` files for the two parts of the library.
One `podspec` with two `subspec`s didn't work, because they are not
supposed to create different modules:
https://github.com/CocoaPods/CocoaPods/issues/3979

To fix CocoaPods' `lint` warnings, a git tag should be created in the
repository first, and the `podspec`s should be updated.

Also, the project file is updated to include the `SimulatorStatusMagic`
library as a subproject instead of the sources directly. This makes it
easier to import the framework in CocoaPods.

Closes https://github.com/zmeyc/UITestUtils/issues/2